Kids criado e tabela do ps melhorada
