+++ defs.h
@@ -124,7 +124,6 @@ int             cps(void);
 int             getusage(int pid);
 int             getpriority(int pid);
 int             setpriority(int pid, int prio);
-int             children(int n);
 
 // swtch.S
 void            swtch(struct context**, struct context*);
