+++ user.h
@@ -27,7 +27,6 @@ int cps(void);
 int getusage(int pid);
 int getpriority(int pid);
 int setpriority(int pid, int prio);
-int children(int n);
 
 // ulib.c
 int stat(const char*, struct stat*);
