+++ usys.S
@@ -33,4 +33,3 @@ SYSCALL(cps)
 SYSCALL(getusage)
 SYSCALL(getpriority)
 SYSCALL(setpriority)
-SYSCALL(children)
